requirements.txt
CommonMark==0.5.4
Django==1.7.1
Pillow==2.6.1
Unidecode==0.04.16
argparse==1.2.1
coverage==3.7.1
django-allauth==0.18.0
django-appconf==0.6
django-bootstrap-toolkit==2.15.0
django-debug-toolbar==1.2.2
django-geoposition==0.2.2
django-tastypie==0.12.1
docutils==0.12
mimeparse==0.1.3
oauthlib==0.7.2
pyflakes==0.8.1
python-dateutil==2.2
python-mimeparse==0.1.4
python-openid==2.2.5
requests==2.4.3
requests-oauthlib==0.4.2
six==1.8.0
sqlparse==0.1.13
wsgiref==0.1.2
